+++ examples/regression/4-tutorials/plot_main-tutorial-regression.py
@@ -40,8 +40,7 @@ from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import PolynomialFeatures
 
 from mapie.metrics import regression_coverage_score
-from mapie.quantile_regression import MapieQuantileRegressor
-from mapie.regression import MapieRegressor
+from mapie.regression import MapieQuantileRegressor, MapieRegressor
 from mapie.subsample import Subsample
 
 os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"