+++ mapie/__init__.py
@@ -4,8 +4,6 @@ from ._version import __version__
 __all__ = [
     "regression",
     "classification",
-    "quantile_regression",
-    "time_series_regression",
     "metrics",
     "__version__"
 ]