+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 ##### (##########)
 ------------------
 
+* Migrate conformity scores classes into conformity_scores module
+* Migrate regression classes into regression module
 * Add split conformal option for regression and classification
 * Update check method for calibration